Move tests to appropriate places
