+++ subprojects/plugins/src/integTest/groovy/org/gradle/groovy/compile/CachedGroovyCompileIntegrationTest.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.groovy.compile
 
-import org.gradle.AbstractCachedCompileIntegrationTest
+import org.gradle.api.tasks.compile.AbstractCachedCompileIntegrationTest
 import org.gradle.test.fixtures.file.TestFile
 
 class CachedGroovyCompileIntegrationTest extends AbstractCachedCompileIntegrationTest {
