+++ subprojects/scala/src/integTest/groovy/org/gradle/scala/ScalaCompileRelocationIntegrationTest.groovy
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.language.scala
+package org.gradle.scala
 
 import org.gradle.integtests.fixtures.AbstractTaskRelocationIntegrationTest
+import spock.lang.Ignore
 
 import static org.gradle.util.JarUtils.jarWithContents
 
+@Ignore("This test does not work yet")
 class ScalaCompileRelocationIntegrationTest extends AbstractTaskRelocationIntegrationTest {
 
     @Override
