+++ subprojects/scala/src/integTest/groovy/org/gradle/scala/compile/CachedScalaCompileIntegrationTest.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.scala.compile
 
-import org.gradle.AbstractCachedCompileIntegrationTest
+import org.gradle.api.tasks.compile.AbstractCachedCompileIntegrationTest
 import org.gradle.test.fixtures.file.TestFile
 
 class CachedScalaCompileIntegrationTest extends AbstractCachedCompileIntegrationTest {
